requirements.txt
aiogram>=3.18.0,<4.0.0
SQLAlchemy>=2.0.38,<3.0.0
pydantic-settings>=2.8.0,<3.0.0
httpx>=0.28.1,<1.0.0
python-dotenv>=1.0.1,<2.0.0
PyMuPDF>=1.25.3,<2.0.0
python-docx>=1.1.2,<2.0.0
openpyxl>=3.1.5,<4.0.0
faster-whisper>=1.1.1,<2.0.0
